+++ packages/types/src/types.ts
@@ -30,11 +30,15 @@ export interface GuidedAnswersQueryPagingOptions {
 
 export type GuidedAnswerTreeSearchHit = GuidedAnswerTree & { SCORE: number };
 
+export type ProductFilter = { PRODUCT: string; COUNT: number };
+
+export type ComponentFilter = { COMPONENT: string; COUNT: number };
+
 export interface GuidedAnswerTreeSearchResult {
     resultSize: number;
     trees: GuidedAnswerTreeSearchHit[];
-    productFilters: { PRODUCT: string; COUNT: number }[];
-    componentFilters: { COMPONENT: string; COUNT: number }[];
+    productFilters: ProductFilter[];
+    componentFilters: ComponentFilter[];
 }
 
 export type GuidedAnswerNodeId = number;
@@ -127,7 +131,7 @@ export const HTML_ENHANCEMENT_DATA_ATTR_MARKER = 'data-guided-answers-command-8d
  * Action types for redux
  */
 export type GuidedAnswerActions =
-    | UpdateGuidedAnserTrees
+    | UpdateGuidedAnswerTrees
     | SelectNode
     | UpdateActiveNode
     | UpdateLoading
@@ -142,10 +146,13 @@ export type GuidedAnswerActions =
     | GuideFeedback
     | SendFeedbackOutcome
     | SendFeedbackComment
-    | BetaFeatures;
+    | BetaFeatures
+    | SetProductFilters
+    | SetComponentFilters
+    | ResetFilters;
 
 export const UPDATE_GUIDED_ANSWER_TREES = 'UPDATE_GUIDED_ANSWER_TREES';
-export interface UpdateGuidedAnserTrees {
+export interface UpdateGuidedAnswerTrees {
     type: typeof UPDATE_GUIDED_ANSWER_TREES;
     payload: GuidedAnswerTreeSearchResult;
 }
@@ -237,4 +244,21 @@ export const SEND_FEEDBACK_COMMENT = 'SEND_FEEDBACK_COMMENT';
 export interface SendFeedbackComment {
     type: typeof SEND_FEEDBACK_COMMENT;
     payload: FeedbackCommentPayload;
+}
+
+export const SET_PRODUCT_FILTERS = 'SET_PRODUCT_FILTERS';
+export interface SetProductFilters {
+    type: typeof SET_PRODUCT_FILTERS;
+    payload: string[];
+}
+
+export const SET_COMPONENT_FILTERS = 'SET_COMPONENT_FILTERS';
+export interface SetComponentFilters {
+    type: typeof SET_COMPONENT_FILTERS;
+    payload: string[];
+}
+
+export const RESET_FILTERS = 'RESET_FILTERS';
+export interface ResetFilters {
+    type: typeof RESET_FILTERS;
 }